+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,18 @@
 
 ### Patch Changes
 
+- ai@6.0.105
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@6.0.104
+
+## 0.0.1
+
+### Patch Changes
+
 - ai@6.0.103
 
 ## 0.0.1